# docs/swagger.yaml
openapi: 3.0.2
info:
    title: Games-API
    version: 1.0.0
    description: >-
        ### This is a Swagger UI for our Games API  \n\n### Error handling\nErrors are presented using the
        HTTP status codes 400-599. The response include a more detailed explanation about the error in the
        body.\n\n```\n{\n  \"error\": \"Cleartext explanation of the error that occurred\"\n}\n```\n\n###
        Generic API errors\nThese error responses can be returned by any of the described endpoints.\n\n| HTTP
        status  |  Description |\n|---|---|\n| 400| Bad input parameters or invalid request. More information
        can be found in the response body.\n| 401| The consumer does not have access to the requested resource
        or the provided token is not valid.\n| 404| The requested resource can not be found or the user does
        not have access to the resource.\n| 5XX| A server error has occurred while processing the request.
paths:
    /games:
        get:
            requestBody:
                description: A new game to be created
                required: true
            tags:
                - games
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                list of games:
                                    value:
                                        - Wicher 3
                                        - Cyberpunk 2077
                    description: application/json
                '201':
                    content:
                        new-game: {}
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                error:
                                    value:
                                        error: some text
                    description: One or all params are missing
            summary: get a list of all games
            description: |
                return a list of all games API
        post:
            requestBody: {}
            tags:
                - new-game
            responses: {}
            summary: Creates a game
            description: Creates a new game
    '/games/{id}':
        get:
            tags:
                - details
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                full details of a given game:
                                    value: |-
                                        {id:1, name: "Wicher 3", price:29.99},
                                        {id:2, name: "Cyberpunc 2077", price:59.99}
                    description: returns full details of a given game
                '404':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                Game /w+ was not found:
                                    value:
                                        - Game \w+ was not found in game array.
                    description: Game was not found in games array
            operationId: id
            summary: get full details of a game
            description: return full details of a given game
        delete:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/delete-game'
            responses:
                '204': {}
                '404': {}
            summary: deliting the game
            description: deleteing the game
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.
        new-game:
            title: Root Type for new-game
            description: ''
            required:
                - price
                - name
            type: object
            properties:
                name:
                    type: string
                price:
                    format: float
                    type: number
            example:
                name: Tetris
                price: 0.99
        delete-game:
            description: ''
            required:
                - id
            type: object
            properties:
                id:
                    description: ''
                    type: string
            example:
                id: 1